fix signed-compare warnings
The m_iSelection array could probably be made unsigned... IF we were
confident we never try to assign it a negative number, of which fact I'm
not convinced.

+++ src/EditMenu.cpp
@@ -398,7 +398,7 @@ void EditMenu::OnRowValueChanged( EditMenuRow row )
 			StepsType orgSel = StepsType_Invalid;
 			if( !m_StepsTypes.empty() ) // Not first run
 			{
-				ASSERT( m_StepsTypes.size() > m_iSelection[ROW_STEPS_TYPE] );
+				ASSERT( (int) m_StepsTypes.size() > m_iSelection[ROW_STEPS_TYPE] );
 				StepsType orgSel = m_StepsTypes[m_iSelection[ROW_STEPS_TYPE]];
 			}
